--- example-dvc-exp-init/.gitattributes
@@ -1,2 +1,0 @@
-*.dvc linguist-language=YAML
-dvc.lock linguist-language=YAML
