--- example-dvc-exp-init/code/.gitignore
@@ -1,6 +1,0 @@
-# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
-
-.venv
