--- example-dvc-exp-init/code/params.yaml
@@ -1,5 +1,0 @@
-train:
-  epochs: 10
-model: 
-  conv_units: 16
-
