--- example-dvc-exp-init/code/src/__init__.py
@@ -1,1 +1,0 @@
-
